+++ src/main/java/com/example/project/demos/web/dto/materialPackageDetail/EditDTO.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -30,7 +31,7 @@ public class EditDTO {
      * 每锅重量
      */
     @ApiModelProperty(value = "每锅重量")
-    private Double packageWeight;
+    private BigDecimal packageWeight;
     /**
      * 创建者
      */
