+++ src/main/java/com/example/project/demos/web/dto/materialPackageDetail/QueryByIdOutDTO.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -34,7 +35,7 @@ public class QueryByIdOutDTO {
      * 每锅重量
      */
     @ApiModelProperty(value = "每锅重量")
-    private Double packageWeight;
+    private BigDecimal packageWeight;
     /**
      * 创建者
      */
